Move observability primitives into services
Observability wrappers and metric helpers now live under a service-owned boundary:\n\n    import "github.com/valon-technologies/gestalt/server/services/observability"\n    import "github.com/valon-technologies/gestalt/server/services/observability/metricutil"\n\n    provider := observability.InstrumentAgentProvider("openai", agentProvider)\n    db := metricutil.InstrumentIndexedDB(indexedDB, "system")\n    attrs := metricutil.RPCAttrs(metricutil.RPCMetricDims{\n        Role:            metricutil.RPCRoleProviderClient,\n        ProviderName:    "openai",\n        HostServiceName: "agents",\n    })\n\nThis is an import-path and ownership move only. Metric names, attributes, span behavior, provider wrappers, IndexedDB instrumentation, meter caching, and gRPC/HTTP telemetry options are unchanged.
