+++ gestaltd/internal/bootstrap/agent_runtime_metrics.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/valon-technologies/gestalt/server/internal/observability"
+	"github.com/valon-technologies/gestalt/server/services/observability"
 	"go.opentelemetry.io/otel/attribute"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
