+++ gestaltd/internal/server/handlers_auth.go
@@ -14,8 +14,8 @@ import (
 	"time"
 
 	"github.com/valon-technologies/gestalt/server/core"
-	"github.com/valon-technologies/gestalt/server/internal/metricutil"
 	"github.com/valon-technologies/gestalt/server/internal/principal"
+	"github.com/valon-technologies/gestalt/server/services/observability/metricutil"
 )
 
 type loginRequest struct {
