+++ gestaltd/services/invocation/catalog.go
@@ -11,8 +11,8 @@ import (
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/core/catalog"
 	"github.com/valon-technologies/gestalt/server/internal/authorization"
-	"github.com/valon-technologies/gestalt/server/internal/observability"
 	"github.com/valon-technologies/gestalt/server/internal/principal"
+	"github.com/valon-technologies/gestalt/server/services/observability"
 	integration "github.com/valon-technologies/gestalt/server/services/plugins/declarative"
 	"go.opentelemetry.io/otel/attribute"
 )
