+++ gestaltd/services/plugins/grpc_telemetry.go
@@ -1,7 +1,7 @@
 package plugins
 
 import (
-	"github.com/valon-technologies/gestalt/server/internal/metricutil"
+	"github.com/valon-technologies/gestalt/server/services/observability/metricutil"
 	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 )
