+++ gestaltd/services/runtimehost/grpc_telemetry.go
@@ -3,7 +3,7 @@ package runtimehost
 import (
 	"strings"
 
-	"github.com/valon-technologies/gestalt/server/internal/metricutil"
+	"github.com/valon-technologies/gestalt/server/services/observability/metricutil"
 	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 )
